Move presets_test content to test.c
